I finished up the installation section of the...
I finished up the installation section of the manual.  As part of this,
I simplified the README file in the main directory.  Frankly, we had
duplicate information and it's just easier to look at the pdf.  Also, I
simplified some of our CMake target names.  Specifically, we use to have
ENABLE_CPP_LIBRARY in order to compile Optizelle.  This has been
simplified to ENABLE_CPP, which matches ENABLE_PYTHON, and
ENABLE_MATLAB.  We used to suffix this flag with _LIBRARY since at one
point we allowed a header only installation.