I modified the third party build setup to...
I modified the third party build setup to install JsonCpp, BLAS, and
LAPACK to the CMAKE_BINARY_DIR.  By doing this, I'm able to then
integrate these libraries into a single install step.  Before, these
third party libraries were implicitely installed after their
compilation.  The affect of this is that it's much easier to use cpack
to grab a working version of the compiled code.  Likely, if you're
having this build setup build these libraries for you, you're going to
need to distribute them as well.  That also simplified some of the
documentation for installing and uninstalling Optizelle.  Finally, I'm
playing funny name games with the documenation to make sure that the
version number of the library is accurately depicted in the file name of
the documentation as well as internally in the manual.  I've done this a
little bit already, but now I'm doing it more.