Another intermediate check-in.
Another intermediate check-in.  I've been working on the basic API
section and I noticed that we're most likely going to want to have a
nice reference manual in addition to the tutorial.  As part of that, we
need to reuse information about the different structures present in
Optizelle.  In order to fascilitate that, I'm using the collect package
inside of LaTeX.  This allows us to make a figure, code, whatever and
then assign it a name that can be reused later.  It's a bit easier to
use than making everything a macro, so I'll try that out for now.  In
addition, I continue to remove all of the old code spcifications and
examples as we clear them out.