I continue to work on the documentation.
I continue to work on the documentation.  Something that has come up is
that it turns out listings works incredibly poorly inside of a long
table.  As such, in order to include longer code examples, I now play a
game with mdframed in order to box our code examples.  Overall, this
looks nice, but I ran into some seriously irritating bugs because
there's some kind of odd reaction with mdframed and the macro
\baselinewidth.  I'm not sure what, but it creates an infinite number of
pages.  As such, beware.  In any case, I'm still futzing with the form
and trying to balance the complexity of the presentation.