I removed the mdframed and collect packages.
I removed the mdframed and collect packages.  Mostly, the mdframed
package kept creating an infinite number of pages, which is bad.  What's
more aggravating, was that I couldn't figure out what was triggering it
easily, which means that my macros were conflicting with it.  Also bad.
In any case, I've moved to a scheme where we basically create fancy
lists of elements in order to list examples, apis, and CMake flags.
It's not quite as pretty as what I was using before, but it seems to be
robust.  As far as the collect package, I think I can get away with
lrboxes for the time being.