I'm continuing to fill out the manual.
I'm continuing to fill out the manual.  As part of that, I have most of
the section on preconditioning complete.  Now, during this discussion, I
realized that my preconditioner for the Rosenbrock function was
incorrect in code.  Basically, I had a scaled version of the inverse.
For our Krylov methods, this doesn't really affect anything; the
eigenvalues still clustered just fine.  For precondioned nonlinear-CG,
this caused all sorts of problems with the change.  In short, I had to
retune all of the parameters and had to do so in some nonintuitive ways.
For example, the algorithm actually takes longer to solve in some
situations with a more accurate line search because the algorithm gets
caught in a bad region.  In theory, this could happen to any algorithm
if we get unlucky, but it seems particularly endemic to nonlinear-CG.
In any case, I have that sorted out, but in doing so I found a bug in
the Matlab interface.  Here, if we were to run the example more than
once, we'd get a segfault.  I traced the problem back to were I was
caching the global instance of Optizelle.  From what I can tell, between
runs, the memory where Optizelle is located moves.  Now, in order to
speed things up, I cached this location, which means that the code was
segfaulting because it couldn't find the Optizelle structure.  In order
to fix this, I now made that pointer global and I refresh it every time
I enter the Optizelle wrapper code.  This is slightly less efficient and
sort of ugly, but it works.