I completed the basic discussion of the...
I completed the basic discussion of the preconditioners in the
documentation.  In order to do this, I modified the simple_equality
example to use a left Schur preconditioner, which it doesn't really
need, but helps illustrate the point.  As part of this, I found that I
had errors in how the bundle of functions were defined in both Matlab
and Python.  Basically, in C++, we have a nice way to telling if the
preconditioner, or some function for that matter, is defined.  Simply,
we look at the pointer and tell if it's null.  In Python and Matlab,
this was not correct.  In Python, the default objects were set to None
and not checked.  This led to a segfault.  In Matlab, I set the default
objects to functions when structs were expected, which led to a
segfault.  In both of these cases, I now define proper objects for
everything, but the default implementation of the functions required in
these objects throws the errors.  These errors are caught properly.
Now, it's a little bit odd since not defined the preconditioner will say
something like the eval function is not defined rather than the
preconditioner is not defined.  For the time being, I'll live with that.
Sometime later, I'll come up with a better scheme.