I removed most of the last remnants of the TRS,...
I removed most of the last remnants of the TRS, trust-region scaling,
operator.  Though, for the time being, I've left it in the linear
algebra solvers.  Originally, I added these operators since the KNITRO
algorithm uses this scaling in the trust-region method along with the
interior point method.  The idea is that we can use information about
how close we are to infeasibility in order to adjust the trust-region.
More specifically, it's basically a scaling of inv(diag(x))diag(z).  In
our case, the situation was always slightly more complicated since we
use a different formulation for the interior point in order to allow
cone constraints.  In addition, we don't put bounds directly on x, but
on h(x), which allows both upper and lower bounds as well as proper
projections from a Hilbert space into something that's finite
dimensional.  Anyway, I had it worked out on how to insert this
operator, but the performance of it was always terrible.  That could be
that I implemented it wrong.  It also could be that it just doesn't work
very well.  In any case, I had it disabled for quite some time, but left
the operators in the code.  Now that I'm working on the documentation
and this operator would have been exposed, it's best to just remove it.
The truncated Krylov methods still allow the operator to be used, so if
I can figure out a way to make it work, I can add it again later.  As
one final note, from my experience, adjusting the inner product is the
best way to modify the trust-region shape, not thse operators.