This is a pretty big check in with respect to...
This is a pretty big check in with respect to modifications of the
manual.  Basically, I've finished up a rough draft of the parameters
section.  At this point, we should have a full description of all of the
paramters in the state, their defaults, and their valid values.  In
order to do this, I rigged up a system to automatically pull in their
default value as well as valid value range from the source.  In a
technical sense, this is sort of neat, but it turns out that it takes
*forever* to compile now.  Honestly, at this point, I'm not sure that
there's anything that I can do about it.  Basically, the listings
package is opening and reopening the file without any sort of caching.
It'd be nice to speed things up, but it's something that I'll worry
about later.  In any case, with any luck, this'll keep the manual more
inline with the code.