I modified the macros in the manual, so it was...
I modified the macros in the manual, so it was easier to add an index.
Now, the index is active and indexes alphabetically most of our
parameters.  Next, I located and fixed a segfault in the json
restart/serialization routines.  We rely on template specialization in
order to add more json serialization routines.  In general, this works
fine, but if we try to serializze and do not have a valid seriailzation
routine, we get odd errors since we don't return a string.  Basically,
we got random data.  Anyway, I added a default implementation that
throws a runtime error in order to alert the user now.  In addition, I
fixed a potential problem in the safeguard against small/bad steps.
Previously, if we took too small of a step, we scaled the step by 0,
which should have caused an exit.  There were two problems with this
approach.  First, if the user defined vector for dx contained a NaN,
we'd calculate the norm wrong.  Second, the scal by 0 process isn't safe
because if the step dx contains a NaN, we don't get the zero vector.
Now, we correctly check of norm_dx is a NaN and if it is, we correctly
set the step to zero with the zero function.  Also, I modified the
trust-region routines to set alpha=1 after we find each step.  The
purpose for this is to make it easier to swtich from a trust-region to a
line-search algorithm.  Previously, this quantity would have been NaN,
since it begins unitialized.  This stopped us from switching from a
trust-region to a line-search algorithm after a restart.  Next, I
removed the old rosenbrock_custom_VS example and added a new
rosenbrock_advanced_api example that subsumes it.  In this new example,
we show the creation of a custom vector space, messaging object, and
restart.  This lets us test more features and will make it easier to
refer to working code that demonstrates these features for the manual.
As part of this, I modified our AddOptizelleJsonTest scripts in order to
accept an extra argument.  We use this in order to create restart
examples.  Finally, I'd forgotten to add the restart read routines to
the Optizelle object in MATLAB.  This has been fixed.