I tweaked the consistency between trust-region...
I tweaked the consistency between trust-region and line-search methods.
Specifically, I now try to keep delta in line with alpha and alpha0 at
each iteration.  My hope is that this will make it easier to switch
between algorithms without having to hand tune the parameters.  In
addition, I added alpha and alpha0 to the output for line-search methods
and delta to the output for trust-region methods.  Finally, I tweaked
the rosenbrock_advanced_api example in order to take advantage of these
new set of features.