I finished the section of the manual that...
I finished the section of the manual that describes the output.  I had
some trouble getting terms with pipes to index correctly, so there are
some hacks in there to make this work.  As it turns out, a pipe is a
reserved character in the LaTeX index package, so watch out.  In
addition, I had a number of bugs in the advanced api codes that I needed
to fix.  Namely, I wasn't ever exercising the serialization routines
properly in Python or Matlab.  On top of that, when the code was fixed,
the Python code was segfaulting because I didn't exit properly in the
Serialization wrapper routines.  I've fixed all of that, so that we
should now exit properly if the serializzation routines are not there.
It works properly if they are there.  In matlab, I made the example a
bit more exotic.  As it turns out, I wasn't actually exercising the new
serialization routines when turned on since the isvector test was
defaulting to the Rm serialization.  That was actually a bad test since
I really wanted to test for isvector and isnumeric.  In any case, I
wrapped the vector in a struct array to make sure that we actually run
the new code.  Finally, I modified some of the output from Optizelle
slightly, KryWhy to KryStop, in order to make the manual line up closer
with the output.