I finished up the restart section and added a...
I finished up the restart section and added a new section with all of
the examples as well as the proper indexing.  In addition, I finally
removed all of the code examples that were not compiled from the manual.
Basically, I hide them inside of the restart unit tests, which make
those unit tests sort of ugly, but it makes sure that all of the code
for things like state initialization actually works.  As it turns out,
there were some minor problems with what I had in the manual and
compiling all of this stuff in these unit tests allowed me to ferret out
some existing, minor bugs.