Shared libraries and memory issues on OS X are...
Shared libraries and memory issues on OS X are going to be the end of
me.  Alright, so this is a big check-in that should, in theory, allow
things to work on OS X.  First, I messed with the code so that we should
be clean with pedantic.  Second, on Unix systems, the suffix for Python
extensions is .so.  Evidently, this is the same for OS X even though
shared libraries have an extension of .dylib.  Now, there's a special
bit of code in the CMake build to deal with that.  On the MATLAB side,
Octave goes haywire on OS X.  Every time I call into Optizelle, the
location of the memory for the Optizelle structure moves.  Now, we fixed
this in the past on Unix systems, but the situation was worse on OS X.
For some reason, we had three versions of the optizelle variable.  If we
grabbed the wrong one, things would crash.  I believe this had to do
with trying to create independent mex files.  Basically, the code in
optizelle.cpp was rolled into an object library and pushed into each of
the mex files.  For some reason, this led to multiple versions of the
optizelle variable, which we consequently always grabbed the wrong one.
In order to fix this, I just roll optizelle.cpp into a shared library
called optizelle_matlab and then link to it.  Of course, this still
isn't good enough on OS X.  The location of the optizelle structure
continues to change each time we call into Optizelle.  This includes
when we call the restart routines from the state manipulator.
Essentially, we call getMin, it gets the optizelle variable.  Later, we
call the state manipulator, which calls write_restart, which gets a new
optizelle variable.  When we return to the getMin code, we have the
wrong optizelle variable.  In order to fix this, I setup a stack where
we keep track of the current optizelle variable.  This is strange and
odd, but it seems to work.  Really, all this mess arises from the lack
of an easy way to just grab the enumerated values out of the structure.
Perhaps there's a better way to do this.  As another note, on OS X,
building and linking BLAS/LAPACK will almost certainly conflict with the
system versions.  In order to fix this, don't set the LD_LIBRARY_PATH,
but DYLD_FALLBACK_LIBRARY_PATH.  Also, on my setup, I end up using mac
ports to grab a set of compilers since the llvm and clang don't have
one.  This means that we need to set the new compiler and all of the
associated utilities in cmake.  When we use cmake to also build jsoncpp
and BLAS/LAPACK, these new options were not passed down to the build.
This meant that we built jsoncpp and BLAS/LAPACK with the wrong
compilers, which led to segfaults.  This has been fixed.  Finally, OS X
is more finnicky with ambiguous types and the jsoncpp library.
Therefore, I've casted more values to eliminate this errors.