Updated dates and versions on primary license
I updated the copyright dates on the primary license along with new
information about who did what.  Basically, with the change over to the
git repository, I can't just refer to the revision numbers as to when
Sandia work stopped and OptimoJoe work began.  As such, I tagged the
last Sandia commit as v1.0.1 and now refer to that in the license file,
which also contains the copyright.