Changed order of the state manipulator in getMin
I changed the order of the arguments for getMin so that the state
manipulator comes last.  Originally, I had things setup the way they
were because the state manipulator was a constant reference.  In
general, I have my constant reference arguments come first and the
arguments that are modified come later.  The problem with this setup is
that, most of the time, the users do not use the state manipulator.
Internally, we always have one, but it's not common for the user to
provide one.  As such, even though it always came after the messaging
object argument, I could never remember that.  Really, it's more of an
add-on, so it makes sense to add it on last when we need it.
Internally, I keep it in the same place, so the only thing that's
changing is where it is in the argument list provided to the user.