Begun full MATLAB compatibility update
In theory, Optizelle is compatible with both MATLAB and Octave.  That
being said, I do most of my testing on MATLAB and it has been some time
since I ran the full suite of unit tests on MATLAB.  As there are some
differences in syntax, basically Octave is more permissive, I've been
fixing this discrepencies.  Something else that's been really irritating
is that I need a pretty different test command for ctest in order to
make the tests work between Octave and MATLAB.  Specifically, I need to
escape the quotes to work with MATLAB, but I can't do this with Octave.
Right now, I have a fix for this, which nicely eliminates the need for
the user to specify the flag for command line execution of a command.
Basically, I check the executable for the word "matlab" and if it exists
I just insert all the correct flags now.  Finally, I don't have
everything fully working, but this is a good start.