Fixed MATLAB CMake configuration
I had a bug where the correct flags were not being sent to MATLAB in
order to run off the command line for ctest.  Basically, add_test
accepts a list of commands.  For whatever reason, I need to have
-nosplash -nodesktop -r in a single string and not as separate strings.
It looks like I had this correct for the test with extra arguments, but
not the regular case.  Second, I eliminated the MATLAB_RUN_FLAG argument
and now automatically detect it.  I detect it by looking for the string
matlab in the MATLAB_EXECUTABLE variable.  In theory, this could break
if someone had a renamed executable.  Hopefully, they won't do that
before I can come up with a better solution.  Finally, I removed the
MATLAB_RUN_FLAG documentation from the manual.