Added more information about MATLAB building
I added a little more information about how to build for MATLAB/Octave.
Basically, the folders for both the library and include are a little
hard to find, so I added more information on where they're included.  In
addition, sometimes on Windows, we have to include libmx in addition to
libmex, so I added that information as well.