Completed a new set of revisions on the manual
I did another run through on the manual correctly several different
errors.  Outside of some wording errors, some of the code imports were
not working correctly, which is why I touch some files liek optizelle.h.
In addition, I finally decided that a default iteration maximum of 10
was more confusing to the users than helpful.  Basically, it should
probably be 1 or infinite.  I decided to go the latter route and now set
the maximum number of iterations to numeric_limits <Natural>::max(),
which is pretty large.  This caused some problems in Python since we use
integers in Python and not longs.  This necessitated some minor updates
to the Python code, but everything seems to be working correctly.
Finally, I tried to add better cross references in the manual as well as
some more precise explanation for some of the derivations.