Modified the LAPACK build procedure
On Windows, CMake by default installs libraries to the bin directory
since windows automatically searches this directory when running an
executable in the same directory.  Unfortunately, this breaks our
Windows builds since we expect the lib directory to contain all the
libraries.  In order to fix this, I added two custom targets that
forcefully empty the bin directory into the lib directory for our
external builds.  I'm a little worried this will go awry if someone does
a highly parallel build on Windows, but hopefully it'll work for now.