Modified the initial inequality multiplier
I'm playing with how we choose the initial inequality multiplier.
Before, we set

z = (<e,e> / <h(x),e>) e.

This insured that mu_est started out at 1.  Really, this made it easy to
see if mu was small and we achieved convergence with respect to the
inequality constraint.  At the same time, I'm not sure if this deals
with scalings properly.  As an alternative, this check-in modifies this
procedure so that we now set

z = ( || h(x) || / || inv(L(h(x))) e || ) inv(L(h(x))) e

This insures that || z || = || h(x) || and that
mu_est = || h(x) ||  / || inv(L(h(x))) e ||.  More importantly, we now
have that h(x) o z = mu_est e on the first iteration, which I believe
means we're on the central path.  In any case, I'm still trying to get a
handle on the performance to see how well this works in practice.