Allowed a more permissive value of alpha0
Recall, alpha0 is the base line-search parameter.  Generally, we want it
to be positive, but there's also a section of the code that sets these
values everytime we set the trust-region.  Basically, we try to keep the
line-search parameters in sync with the trust-region and vice versa so
that it's easier to switch between algorithms.  In any case, when we get
a bad Hessian, it's possible that we set delta to 0.  This happens
because dx=0 and then we set delta to ||dx||/2.  This means that we set
alpha0 to zero, which then causes an issue with the restart mechanisms
since alpha0 previously needed to be positive.  This issue has now been
fixed.