Updated SR1 algorithm and added safeguards
I modified the SR1 algorithm to use a numerically more stable
computation.  Basically, we should do the inner products last and we
were previously doing them first.  In addition, I added a safeguard to
the algorithm that had some divide by zero problems.  Basically, in SR1,
we eventually have to divide by <si,yi-Bisi>.  There are a couple of
things that can go wrong here.  First, si could be zero, but we caught
that previously since si=xi-xim1=dx and we have a stopping critieria
that looks for small dx.  Second, si could be orthogonal to yi-Bisi.  In
order to combat this, we now use the scheme suggested by Nocedal and
Wright where we require

        <si,yi-Bisi> > epsilon ||si|| ||yi-Bisi||

where epsilon is equal to the square root of machine precision.
Finally, we could have yi-Bisi=0.  Generally, for unconstrained
problems, this doesn't come up.  Basically, if the problem is quadratic
with m variables, then SR1 gives a perfect approximation after m
interations.  This mean that after m iterations, we have that yi-Bisi=0.
Now, for unconstrained problems, we exit the optimization by then, so
there's no problem.  However, for inequality constrained problems, we
have to wait until we drive down mu to something small.  This means that
we don't necessessarily exit after m iterations, which means that
yi-Bisi can become small.  In order to combat this, I implemented a new
procedure that requires

        | s'(y-Bs) | > epsilson max_i | si'(yi-Bsi) |

Essentially, we want to guarantee that the new information we're adding
about the Hessian is relavant when compared to the information already
in the SR1 operator.