Updated unit tests to reflect algorithmic changes
I updated the unit tests to reflect the changes made to both the
initial choice of inequality multiplier as well as SR1 algorithm.  For
the most part, the problems because more robust and the iteration count
went down.  This was especially pronounced for Newton-CG with a
backtracking line-search.  I did have some issues with the predictor
corrector algorithm, which I'm not happy about.  Honestly, it seems a
bit flakey right now, so I need to spend some time figuring out why the
algorithm performs in the way it does.  By flakey, for a particular
value of gamma, it may or may not converge and it took way too much
tweaking to get something that performed well.