Added iter and name as arguments to serialize
I modified the API so that the serialize routine in Serialization now
provides access to the name of the variable, as a string, and the
current iteration.  The reason for this change is that it can be
difficult to figure out an appropriate file name in case we need to
write a restart file to a binary file and then only include that file
information in the json file.  Especially for big problems, writing to
external binary files is essential.  But, yes, previously we just gave
the serialize routine a vector and it was never clear, at least to me,
as to where to write this information.