Added example that demonstrates restarts to file
I added a new example that demonstrates how to write restart files to
file.  Specifically, how to write the variables to file and not directly
to json, which is a common use case for large scale optimization
problems.  Basically, this demonstrates the functionality added earlier
in this tree.

In the process of doing this, I uncovered a few different bugs.  First,
it turns that that mu_typ was not being written out the restart
packages.  This has now been fixed.  Next, there was a more subtle
problem with the StateManipulator being called at the
EndOfOptimizationIteration iteration.  Basically, this is the best place
to write the restart file.  However, this is also the location where the
stopping conditions were modified.  Since the user's state manipulator
is always called first, the restart file was being written with the
wrong stopping condition.  If we recall, the state manipulators that
implement the inequality and equality constraints adjust the stopping
conditions here.  In order to fix this, I needed to add another location
for the state manipulator called AfterCheckStop.  This is a better place
to adjust the stopping conditions.  Then, EndOfOptimizationIteration
becomes a safe place to write the restart files.  Now, added to our
enumerated types requires changes to the MATLAB/Octave and Python
interfaces, so those were modified as well.  Finally, there were some
minor bugs in rosenbrock_advanced_api, which became exposed when dealing
with our new example.  These have also been fixed.