Changed the reporting of ||dx|| during LS algos
I modified the meaning of ||dx|| during a line-search algorithm.
Before, I was trying to process the step so that ||dx|| was the size of
the raw step, not scaled by alpha.  The problem with this was that after
a rejected line-search step, I was misreporting what the norm was
because I wasn't scaling it properly before we hit the state
manipulator.  Now, I could have fixed this, but in some sense this
probably wasn't the way to go from the start.  When working on a
problem, I find it much easier to tune the step size if I know what the
size of || alpha dx || is, rather than ||dx||.  In the latter case, I
just end up multiplying things anyone.  It also simplifies our code.