Playing with rescalings and starting dual vars
I'm in the middle of playing with different ways to rescale the problem
and different ways to choose the initial dual variable.  Originally, I
tried adding a rescaling of the form

A = h'(x)* L(h(x)) h'(x)

to the Newton system

H(x) dx = -grad f(x)

Bascally, that becomes

A* H(x) A dx = -A* grad f(x)

My hope was that I would be able to basically zero out the elements of
the gradient corresponding to the elements on the boundary.  In
practice, this worked poorly because the introduction of two of the A
operators in the linear sytem caused severe numerical problems.
Basically, we're inverting to A's.  As such, it appeared to behave
poorly.

Really, if we think about the current gradient calculation wih the
interior point pieces

grad f(x) - mu h'(x)* inv(L(h(x))) e

This kind of does the same thing.  If something is near the boundary, we
redirect the gradient in the other direction.  Well, we do if mu is
large enough.  As such, I added code to let the user choose an initial
mu more freely.  This led to an addition of news ways to choose the
initial dual variable.  Right now, I'm playing with setting

z = (mu <e,e> / <h(x),h(x)> ) h(x)

since the Hessian modification, in the case of bound constraints on x
becomes

inv(L(h(x)) L(z) = (mu <e,e> / <h(x),h(x)>) I

and the interior point estimate is

<h(x),z>/<e,e> = mu

Honestly, I don't think it works that well, but it's something to play
with.