Removed scaling and cleaned up new parameters
I removed the scaling from the previous check in and then cleaned up the
code related to some new diagnostic parameters.  Basically, in an
interior point method, we often shorten our steps to insure feasbility
with respect to the inequality constraints.  Before, all of this was
done internally.  Now, I added those parameters to the state and display
them when the msg_level is set to 2 or higher.  I did add these
parameters to the C++ code in the last check-in, but this check-in
cleans all of it up and adds it to MATLAB/Octave and Python.  I also
changed the choice of the dual variable back to what it was before in
the develop branch.  As such, all of the current unit tests should be
passing.