Added a scaled variant of an interior point method
I added sort of an odd, scaled interior point method as an experiment.
Basically, our complementary slackness condition is normally relaxed
into

h(x) o z = mu e

Instead, I relaxed it into

h(x) o z = mu h(x) o h(x)

This gives a gradient modification of

- mu h'(x)* h(x)

and a Hessian modification of

h'(x)* inv(L(h(x))) (
    h'(x)dx o z - mu h'(x)dx o h(x) - mu h(x) o h'(x)dx)

with a merit function modification of

-(mu/2) < h(x), h(x) >

Overall, it only works ok and I'm not fully convinced that it works with
first order algorithms.  The idea behind it was to create a situation
where z was encouraged to become mu h(x).  This helps moderate the
conditioning of the Hessian modification.  In addition, if we look at
the contour plots of

x y - mu = 0

we see that the graph approaches 0 as either x or y gets large.  For
poorly scaled problems, this feels like it causes an issue.  Conversely,

x y - mu x^2 = 0

asks y to also get large if x is large, which seems like it should work
better for poorly scaled systems.

In any case, I'm not all that happy with the performance, but I thought
the code was interesting enough to check in.