Changed our initial z to be mu inv(L(x)) e
I change our initial inequality multiplier to be

z =  mu inv(L(x)) e

where mu is chosen by the user and has a default of 1.0.  In this way,

h(x) o z = mu e

and

mu_est = < h(x) , z > / < e , e > = mu

Technically, I had this code floating around from some previous changes,
but I just reenabled it.  Really, by messing with the initial mu, we can
force the step to be more or less internal to the problem.  This
provides an ability to tweak a problem to force things to work in many
cases.

In addition, I cleaned up the code regarding alpha_x and alpha_z.  Now,
it correctly outputs to our diagnostics.  Specifically, we only see
alpha_z when doing a PrimalDual interior point method and not with the
other methods where we ignore this parameter.

I also updated the manual and correctly document the changes to z,
alpha_x, and alpha_z.