Added SQL cone constraint support to MATLAB
I added basic support for SQL cone constraints to MATLAB.  This has
existed for sometime in the C++ code.  Now, that being said, there are
still some gaps.  I've not completed the serialization routines for this
problem yet.  Part of the reason is that the SQL code in C++ uses a
different way to store the information that MATLAB.  Basically, in C++
it's convenient to have everything in memory laid out in a columnwise
vector since we're going to pass the operations to LAPACK anyway.
However, in MATLAB, this is a pain to deal with because we have to
constantly reshape our information.  Therefore, the serialization for
the routines would look different, which goes against our general theme
of letting someone take a serialized problem from one language and move
it to another.  As such, this will require some more thinking.  In
addition, the behavior of the implementations is slightly different due
to different implemenations of the operators.  Since we're in MATLAB, we
can just use things like eigs for the generalized eigenvalue problem,
which will give a different behavior that the hand coded one in C++.