Fix an error with the MATLAB error handling
In the MATLAB error handing, there's an issue that comes up with MATLAB
and not Octave.  Basically, MATLAB doesn't correctly call the error
function in the Messaging object.  Now, we just end up calling the
MATLAB error function anyway, so for the time being let me disable this.
This whole system is going to be overhauled anyway.