Fixed a bug where MATLAB would for iter_max to 0
I fixed a bug where MATLAB/Octave would force iter_max to 0.  The
default value for iter_max is std::numeric_limits <Natural>::max() from
C++.  Basically, the largest unsigned integer that's architecture
dependent (size_t).  Now, MATLAB/Octave can handle integers, sort of,
but it really just works with doubles most of the time.  As such, we end
up casting these doubles to integers.  This generally works well, but
there was an issue when reading in a MATLAB/Octave state or reading in a
MATLAB/Octave restart package that this would overflow.  Then, iter_max
would be set to 0, which caused all sorts of problems.  The fix is to
cast ::max() and ::min() to doubles, compare to what C++ pulls from
MATLAB/Octave, and then hard set to ::min() or ::max() if we detect an
under or overflow respectively.

This fixes #36.