Cleaned up caching and unconstrained code
I cleaned up the caching and the unconstrained code.  Basically, as much
as I would like MATLAB/Octave to allow real functional programming,
alas, it is not the case.  Before, I was introducing variables into my
functions by doing an abstraction followed by an application.  This was
because MATLAB/Octave doesn't have a "let x = e" introduction statement
in its grammar.  This made it hard for me to modify, so I just removed
all of that and added new functions.  I also went through and changed
the caching to just use isequal.  This is more accurate and prevents
problems with poorly scaled variables.