Reduced Hess-vec product to 1-forward 1-adjoint
I reduced the number of solves used by the Hessian-vector product, in
the reduced space method, down to 1-forward and 1-adjoint solve.
Basically, this halves the cost to what we had before.  In order to
accomplish this, I cache things differently and introduce a new operator

psi(x) = inv(g'_2(x,phi(x))*) grad_2 f(x,phi(x))

What this does is cache a value used in the gradient solve that can be
reused in the Hessian-vector product.  In addition, I combine all of the
solves for the Hessian-vector product that involve phi'(x)*.  This meant
that I needed to split up phi''(x).  Although it sounds convoluted, it's
pretty straight forward.  Basically, just write out all of the equations
for the gradient and Hessian-vector product and then notice that the
stuff in psi can be cached and that we can factor out a phi'(x)*.  In
any case, this code should be much faster than what we had before.