Added unit test for the parest code
I added a unit test for the parest, parameter estimation, example code.
Since we generate random problems, we don't validate any of the results.
Hence, we just run it.  Nevertheless, it exercises a good bit of
Optizelle, so it should be a test.

By the way, we now ignore any directories in the primary directory with
the name build*.  Before, we just ignored build.  This was irritating me
since I have multiple build directories.