Added the parametrization function
I added a parametrization function, which is in some sense dual to the
interpolation function.  Basically, the interpolant has the form

f(alpha,A,b)(x) = alpha'*phi(A*x+b)

and the parametrization has the form

f(x)(alpha,A,b) = alpha'*phi(A*x+b)

Basically, one is in terms of the points and the other is in terms of
the parameters.  Note, this function is vectorized, but not quite as
efficiently has the interpolation function, so there's likely some room
for improvement.