Modified example to match MATLAB toolbox
I modified the example to more closely match how the MATLAB neural net
toolbox works.  Basically, MATLAB always using scaling operators both on
the domain and codomain of the problem.  In addition, even in a single
layer MLP, it has an output bias weight, which I'm pretty sure is not
needed by the theory.  Anyway, I may futz with this later, but for the
time being its nice to be able to take a MLP from MATLAB and use some of
these routines.