Modified orthogonality check for truncated CG
I modified the orthogonality check for truncated CG, which is useful
when running in the equality and constrained case.  Basically, I had
this odd scheme before where everything was stored in a list in a
roundabout way.  Now, we just store the matrix for the check columwise
in vectors.  I also tried to clean up the documentation because I kept
forgetting exactly how and why the check works.