Added a new diagnostics for the augmented solves
I added a new set of diagnostics for the augmented systems solves.
Specifically, I added information about the number of solves, total
number of solves, error in the solve, and desired error in the solve to
the diagnostic output to all of the augmented system solves.  This
should help make it easier to determine when the preconditioner is
failing or something funny is going on.  In addition, it makes assessing
the efficiency much easier since we can just add up the total number of
solves now.  Recall, each one of these solves probably should use a
preconditioner and that preconditioner can have a pretty hefty cost.

Anyway, there's something like 20 new variables that I had to add to the
state.  In theory, I did this correctly for all the C++ with the
restarts and the like.  I have not yet added them to the structures for
the Python and MATLAB/Octave interfaces.  I also haven't added them to
the manual, but I think that the documentation stubs for pulling in
information from listings is there.