Added total number of augmented system solves
I added the total number of augmented system solves to the output.  In
theory, we could just add the numbers ourselves, but computers are much
better at that, so now we have a new variable.  I also fixed a bug with
the headers when the message level was 2 and messed around the with the
column headers.