Added total number of krylov iterations
I added the total number of Krylov iterations to the level 3 output.  In
the past, we used this level of output to have the Krylov solver
interleave output into the optimization solves.  This sort of worked.
Sort of.  Ultimately, I turned it off because it was irritating.
Anyway, the variable krylov_iter_total, which still existed, got lost in
this mix.  With this change, it gets reintroduced.  In the long run,
we're probably better off with a different diagnostic output method, but
that'll come later.